Added auto-discovery of GPUs for Nvidia GPU support.
This commit also includes updates to the tests that would start
failing once GPU autodiscovery was turned on. Since GPUs are now a
first class resource, we have to explicilty list them in resource
vectors that attempt to set resources for the whole machine and then
verify that they get the same vector back in an offer.